fix T-gMark scripts

+++ tests/benchmarks/T-gMark/create_graph_op_queries.py
@@ -23,7 +23,7 @@ class gMark():
         # 更新的节点id
         self._update_node_lists = []
         # 每一个类别create操作的个数上限
-        self._create_node_num = self._max_create_op / 2
+        self._create_node_num = int(self._max_create_op / 2)
         self._create_edge_num = self._max_create_op - self._create_node_num
 
     ###########help function###########
@@ -50,7 +50,7 @@ class gMark():
         with open(file_path) as f1:
             reader = csv.reader(f1, delimiter='|')
             for row in reader:
-                write_lists.append(row[1])
+                write_lists.append(row[0])
         f1.close()
         return write_lists
 
@@ -86,7 +86,7 @@ class gMark():
             update_type_lists = []
             id_lists = self.get_csv_nodes(csv_path)
             lens = len(id_lists)
-            print(lens)
+            # print(lens)
             left = np.uint64(1)
             right = np.uint64(lens - 1)  # 文件的行数
             selected_lists = self.Zipf(1.1, left, right, self._max_update_op)
@@ -135,21 +135,23 @@ class gMark():
         return tgdb_write_lists, tgql_write_lists
 
     ######## add query #########
-    def add_node():
+    def add_node(self):
         self._id_to_be = self._id_to_be + 1
         query=f"Create (n:Node {{id:{self._id_to_be}}})"
         tgql_query=f"Create (n:Object {{ title:'Node',id:{self._id_to_be},validTimeStart:timestamp(),validTimeEnd:9223372036854775000 }}) Create (n)-[:OBJECT_ATTRIBUTE]->(att:Attribute {{title:'id',validTimeStart:timestamp(),validTimeEnd:9223372036854775000}})  Create (att)-[:ATTRIBUTE_VALUE]->(:Value {{title:0,validTimeStart:timestamp(),validTimeEnd:9223372036854775000}})  Create (n)-[:OBJECT_ATTRIBUTE]->(att2:Attribute {{title:'name',validTimeStart:timestamp(),validTimeEnd:9223372036854775000}})  Create (att2)-[:ATTRIBUTE_VALUE]->(:Value {{title:0,validTimeStart:timestamp(),validTimeEnd:9223372036854775000}}) ;"
         return query,tgql_query
-    def add_edge(from_node, edge_type, to_node):
+    def add_edge(self,from_node, edge_type, to_node):
         query=f"MATCH (n1 {{id:{from_node}}}),(n2 {{id:{to_node}}}) with n1, n2 Create (n1)-[:{edge_type}]-(n2);"
         tgql_query=f"MATCH (o1:Object)-[o:OBJECT_ATTRIBUTE]-(att {{title:'id'}})-[a:ATTRIBUTE_VALUE]-(v {{title:{from_node}}}),(o2:Object)-[:OBJECT_ATTRIBUTE]-(att2 {{title:'id'}})-[a2:ATTRIBUTE_VALUE]-(v2 {{title:{to_node}}}) with o1, o2 Create (o1)-[:edge_type]-(o2);"
         return query,tgql_query
 
     def _create(self):
-        node_edge_list_1=random.sample(self._update_node_lists,self._create_edge_num)
-        node_edge_list_2=random.sample(self._update_node_lists,self._create_edge_num)
         cypher_lists = []
         tgql_cypher_lists = []
+        if self._max_create_op == 0:
+            return cypher_lists, tgql_cypher_lists
+        node_edge_list_1=random.sample(self._update_node_lists,self._create_edge_num)
+        node_edge_list_2=random.sample(self._update_node_lists,self._create_edge_num)
         for i in range(self._create_node_num):
             create_query,tgql_create_query=self.add_node()
             cypher_lists.append(create_query)
@@ -161,21 +163,23 @@ class gMark():
         return cypher_lists,tgql_cypher_lists
 
     ######### delete query ############
-    def delete_node(node_label,v_id):
+    def delete_node(self,node_label,v_id):
         query=f"MATCH (c:Node {{id: '{v_id}'}})-[e]-() delete e, c;"
         tgql_query1=f"MATCH (p:Object {{id:{node_label}{v_id}}})-[o:OBJECT_ATTRIBUTE]-(att)-[a:ATTRIBUTE_VALUE]-(v) delete o,a,att,v;"
         tgql_query2=f"MATCH (p:Object {{id:{node_label}{v_id}}})-[e]-() delete e,p;"
         return query, tgql_query1,tgql_query2
 
-    def delete_edge(from_node, edge_type, to_node):
+    def delete_edge(self,from_node, edge_type, to_node):
         query=f"MATCH (n1 {{id:{from_node}}})-[e:{edge_type}]-(n2 {{id:{to_node}}}) delete e;"
         tgql_query=f"MATCH (o1:Object)-[o:OBJECT_ATTRIBUTE]-(att {{title:'id'}})-[a:ATTRIBUTE_VALUE]-(v {{title:{from_node}}}),(o2:Object)-[:OBJECT_ATTRIBUTE]-(att2 {{title:'id'}})-[a2:ATTRIBUTE_VALUE]-(v2 {{title:{to_node}}}) with o1, o2 MATCH (o1)-[e:edge_type]-(o2) delete e;"
         return query,tgql_query
 
     def _delete(self):
-        node_edge_list=random.sample(self._update_node_lists,self._create_node_num)
         cypher_lists = []
         tgql_cypher_lists = []
+        if self._max_delete_op == 0:
+            return cypher_lists, tgql_cypher_lists
+        node_edge_list=random.sample(self._update_node_lists,self._create_node_num)
         for to_do in node_edge_list:
             delete_query,tgql_delete_query_1, tgql_delete_query_2=self.delete_node("Node",to_do)
             cypher_lists.append(delete_query)
@@ -189,6 +193,7 @@ class gMark():
         TGQL_write_lists = []
         # 1. get update vertex id lists 2.get queries
         update_id_lists=self._get_type_lists(f"{args.write_path}peak_vertices/update_Num{gmark._max_graph_op}", args.gmark_dataset_path)
+        self._update_node_lists = update_id_lists
         # 2.1 prepare for update query
         cypher_lists, tgql_cypher_lists = self._update(update_id_lists)
         write_lists += cypher_lists
@@ -211,16 +216,16 @@ if __name__ == "__main__":
         description="AeonG create graph operation queries.",
         formatter_class=argparse.ArgumentDefaultsHelpFormatter)
     parser.add_argument("--size",
-                        default="small",
+                        default="sf1",
                         help="Dataset size")
-    parser.add_argument("--num-op",
+    parser.add_argument("--num-op", type=int,
                         default=80000,
                         help="The number of graph operation queries")
-    parser.add_argument("--update-ratio",
-                        default=0.8,
+    parser.add_argument("--update-ratio", type=float,
+                        default=1,
                         help="The update ratio of graph operation queries")
-    parser.add_argument("--delete-ratio",
-                        default=0.1,
+    parser.add_argument("--delete-ratio", type=float,
+                        default=0,
                         help="The delete ratio of graph operation queries")
     parser.add_argument("--write-path",
                         default="../../results/",
@@ -237,6 +242,6 @@ if __name__ == "__main__":
     gmark = gMark(size=args.size, max_op=args.num_op)
     ##### graph operatios generation #####
     write_lists, tgql_write_lists=gmark.get_all_queries(dataset_path=args.gmark_dataset_path, write_path=args.write_path)
-    gmark.write_to_file(f"{args.write_path}graph_op/cypher.txt", tgdb_write_lists)
+    gmark.write_to_file(f"{args.write_path}graph_op/cypher.txt", write_lists)
     gmark.write_to_file(f"{args.write_path}graph_op/TGQL_cypher.txt", tgql_write_lists)
     print("=========done!========")
